AI-Voice-Agent-main/requirements.txt
Flask
twilio
requests
python-dotenv
gradio_client
